Eliminada la actividad de HomeActivity antigua y renombramiento de la nueva (Home2Activity) a HomeActivity.
