+++ app/src/main/java/io/github/jokoframework/activity/LineChartActivity.java
@@ -128,7 +128,7 @@ public class LineChartActivity extends Activity {
     }
 
     private void backToHome() {
-        Intent intent = new Intent(this, Home2Activity.class);
+        Intent intent = new Intent(this, HomeActivity.class);
         startActivity(intent);
         finish();
         overridePendingTransition(R.anim.left_to_right, R.anim.right_to_left);
