+++ app/src/main/java/io/github/jokoframework/activity/MultipleLineChartActivity.java
@@ -211,7 +211,7 @@ public class MultipleLineChartActivity extends Activity {
     }
 
     private void backToHome() {
-        Intent intent = new Intent(this, Home2Activity.class);
+        Intent intent = new Intent(this, HomeActivity.class);
         startActivity(intent);
         finish();
         overridePendingTransition(R.anim.left_to_right, R.anim.right_to_left);
