--- app/src/main/res/layout/activity_home.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.drawerlayout.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/drawer_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="io.github.jokoframework.activity.HomeActivity">
-    
-    <fragment
-        android:id="@+id/imageFragment"
-        android:name="io.github.jokoframework.activity.HomeActivity$MainViewFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:layout="@layout/activity_show_info" />
-
-    <fragment
-        android:id="@+id/menuFragment"
-        android:name="io.github.jokoframework.fragment.NavigationDrawerFragment"
-        android:layout_width="@dimen/navigation_drawer_width"
-        android:layout_height="match_parent"
-        android:layout_gravity="start"
-        tools:layout="@layout/fragment_navigation_drawer" />
-
-</androidx.drawerlayout.widget.DrawerLayout>
